Organize each module more
