+++ api/src/core/auth/exceptions/auth.exception.ts
