+++ api/src/core/auth/local/local.strategy.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
 
-import { Admin } from '@api/features/admin/admin.entity';
+import { Admin } from '@api/features/admin/entities/admin.entity';
 
-import { AuthService } from '../auth.service';
+import { AuthService } from '../services/auth.service';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
