+++ api/src/core/database/postgres-error-codes.enum.ts
@@ -1,4 +1,4 @@
-export enum PostgresErrorCode {
+export enum PostgresErrorCodes {
     NOT_NULL_VIOLATION = '23502',
     FOREIGN_KEY_VIOLATION = '23503',
     UNIQUE_VIOLATION = '23505'
