+++ api/src/core/http/exceptions/http.exception.ts
