+++ api/src/features/admin/admin.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { Admin } from './admin.entity';
-import { AdminService } from './admin.service';
+import { Admin } from './entities/admin.entity';
+
+import { AdminService } from './services/admin.service';
 
 @Module({
     imports: [
