+++ api/src/features/admin/exceptions/admin.exception.ts
