+++ api/src/features/api/api.module.ts
@@ -4,7 +4,7 @@ import { BlogModule } from '@api/features/blog/blog.module';
 import { ProfileModule } from '@api/features/profile/profile.module';
 import { ProjectModule } from '@api/features/project/project.module';
 
-import { ApiController } from './api.controller';
+import { ApiController } from './controllers/api.controller';
 
 @Module({
     imports: [
