+++ api/src/features/api/controllers/api.controller.ts
@@ -3,12 +3,12 @@ import { Controller, Get, HttpCode, Param, Query, Req } from '@nestjs/common';
 import { Request } from 'express';
 
 import { BlogPostService } from '@api/features/blog/services/blog-post.service';
-import { ProfileService } from '@api/features/profile/profile.service';
-import { ProjectService } from '@api/features/project/project.service';
+import { ProfileService } from '@api/features/profile/services/profile.service';
+import { ProjectService } from '@api/features/project/services/project.service';
 
-import { ActiveProfileWasNotFoundException } from '@api/features/profile/profile.exception';
+import { ActiveProfileWasNotFoundException } from '@api/features/profile/exceptions/profile.exception';
 import { BlogPostsWereNotFoundException } from '@api/features/blog/exceptions/blog-post.exception'
-import { ProjectsWereNotFoundException } from '../project/project.exception';
+import { ProjectsWereNotFoundException } from '@api/features/project/exceptions/project.exception';
 
 @Controller()
 export class ApiController {
