+++ api/src/features/blog/services/blog-post.service.ts
@@ -3,8 +3,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 
 import { Repository } from 'typeorm';
 
-import { PostgresErrorCode } from '@api/core/database/postgres-error-code.enum';
-import { InternalServerErrorException } from '@api/core/http/http.exception';
+import { PostgresErrorCodes } from '@api/core/database/postgres-error-codes.enum';
+import { InternalServerErrorException } from '@api/core/http/exceptions/http.exception';
 import { BlogPostAlreadyExistsException } from '../exceptions/blog-post.exception';
 
 import { BlogPost } from '../entities/blog-post.entity';
@@ -28,7 +28,7 @@ export class BlogPostService {
 
         return await this.blogPostRepository.save(post)
             .catch((error) => {
-                if(error.code === PostgresErrorCode.UNIQUE_VIOLATION) {
+                if(error.code === PostgresErrorCodes.UNIQUE_VIOLATION) {
                     throw new BlogPostAlreadyExistsException();
                 } else {
                     throw new InternalServerErrorException();
