+++ api/src/features/profile/exceptions/profile.exception.ts
